+++ src/main/java/ProjectEuler/Problem10.java
@@ -2,11 +2,11 @@ package ProjectEuler;
 
 public class Problem10 implements Solvable {
   public String solve() {
-    long answer = 2L;
+    long answer = 2;
 
     for (int i = 3; i < 2000000; i += 2) {
       if (Helpers.isPrime(i)) {
-        answer += (long) i;
+        answer += i;
       }
     }
